Server reorganized + add games modal window
