+++ assets/components/AppComponent.html
@@ -4,4 +4,5 @@
 		<div component="game-container" id="game-container" class="col-sm-8 col-lg-10 col-sm-offset-4 col-lg-offset-2">
 		</div>
 	</div>
+	<div component="add-games-modal"></div>
 </div>
