+++ assets/components/SidebarComponent.html
@@ -1,6 +1,6 @@
-<ul id="potential-games-list" class="games-list">
-	<li>{{ loading }}</li>
-</ul>
+<div id="potential-games-area">
+	{{ loading }}
+</div>
 <hr>
 <ul id="playable-games-list" class="games-list">
 	<li>No added games.</li>
