+++ assets/sass/_sideBar.scss
@@ -6,6 +6,11 @@
 	width: 16%;
 	overflow-y: auto;
 
+	div#potential-games-area {
+		text-align: center;
+		padding: 20px;
+	}
+
 	hr {
 		margin: 0;
 		border-top: 1px solid #565149;
@@ -16,6 +21,7 @@
 		margin: 0;
 
 		li {
+			color: #A5A5A5;
 			display: block;
 			padding: 10px 20px;
