+++ assets/ts/client/VitrineClient.ts
@@ -1,5 +1,4 @@
 import { ipcRenderer } from 'electron';
-import * as truncate from 'truncate';
 
 import { GamesCollection } from '../models/GamesCollection';
 import { PotentialGame } from '../models/PotentialGame';
@@ -24,7 +23,7 @@ export class VitrineClient {
 		ipcRenderer.send('client.ready');
 	}
 
-	public launchEvents() {
+	public registerEvents() {
 		// TODO: Remove this events pipeline
 		ipcRenderer.on('server.send-game', (event, game) => {
 			$('#game-cover-container').css({
@@ -43,22 +42,18 @@ export class VitrineClient {
 				});
 			}
 		});
-
 		ipcRenderer.on('server.send-game-error', (event, error) => {
 			$('#game-title').html(error);
 			throw new Error(error);
 		});
-
 		ipcRenderer.on('server.add-potential-games', (event, games) => {
 			this.potentialGames.games = games;
 			this.renderPotentialGames();
 		});
-
 		ipcRenderer.on('server.add-playable-games', (event, games) => {
 			this.playableGames.games = games;
 			this.renderPlayableGames();
 		});
-
 		ipcRenderer.on('server.remove-potential-game', (event, gameId) => {
 			this.potentialGames.removeGame(gameId, (error, game: PotentialGame) => {
 				if (error)
@@ -68,12 +63,10 @@ export class VitrineClient {
 				}
 			})
 		});
-
 		ipcRenderer.on('server.add-playable-game', (event, playableGame) => {
 			this.playableGames.addGame(playableGame);
 			this.renderPlayableGames();
 		});
-
 		ipcRenderer.on('server.stop-game', () => {
 			console.log('Game stopped.');
 			this.gameLaunched = false;
@@ -81,20 +74,10 @@ export class VitrineClient {
 	}
 
 	private renderPotentialGames() {
-		$('#potential-games-list').html('');
-
-		this.potentialGames.sort();
-
-		let counter: number = 0;
-		this.potentialGames.forEach((potentialGame: PotentialGame) => {
-			let html: string = '<li game-id="' + potentialGame.uuid + '" class="add-game-link">' +
-				languageInstance.replaceJs('addGame', truncate(potentialGame.name, 35)) +
-				'</li>';
-			$('#potential-games-list').append(html);
-			counter++;
-			if (counter == this.potentialGames.games.length)
-				this.createGameClickEvents(false);
-		});
+		let html: string = '<button id="add-potential-games-btn" class="btn btn-primary" data-toggle="modal" data-target="#add-games-modal">' +
+			languageInstance.replaceJs('potentialGamesAdd', this.potentialGames.games.length) +
+			'</button>';
+		$('#potential-games-area').html(html);
 	}
 
 	private renderPlayableGames() {
@@ -108,64 +91,54 @@ export class VitrineClient {
 			$('#playable-games-list').append(html);
 			counter++;
 			if (counter == this.playableGames.games.length)
-				this.createGameClickEvents(true);
+				this.eventPlayableGames();
 		});
 	}
 
-	private createGameClickEvents(treatingPlayableGames: boolean) {
-		if (treatingPlayableGames) {
-			$('li.play-game-link[game-id]').each((index, value) => {
-				$(value).click(() => {
-					let gameId: string = $(value).attr('game-id');
-					this.playableGames.getGame(gameId, (error, game) => {
-						if (error)
-							throw new Error(error);
-						if (this.clickedGame && this.clickedGame.uuid === gameId)
-							return;
+	private eventPlayableGames() {
+		$('li.play-game-link[game-id]').each((index, value) => {
+			$(value).click(() => {
+				let gameId: string = $(value).attr('game-id');
+				this.playableGames.getGame(gameId, (error, game) => {
+					if (error)
+						throw new Error(error);
+					if (this.clickedGame && this.clickedGame.uuid === gameId)
+						return;
 
-						if (this.clickedGame)
-							$('li.play-game-link[game-id=' + this.clickedGame.uuid + ']').removeClass('selected-game');
-						$(value).addClass('selected-game');
+					if (this.clickedGame)
+						$('li.play-game-link[game-id=' + this.clickedGame.uuid + ']').removeClass('selected-game');
+					$(value).addClass('selected-game');
 
-						let gameCover: string = 'url(' + game.details.cover.split('\\').join('\\\\') + ')';
-						let gameBgScreen: string = 'url(' + game.details.backgroundScreen.split('\\').join('\\\\') + ')';
-						$('#game-cover-container').css({
-							'display': 'block'
-						});
-						$('#game-title').html(game.name);
-						$('#game-desc').addClass('game-desc').html(game.details.summary);
-						$('#game-cover-image').css({
-							'background-image': gameCover,
-							'background-repeat': 'no-repeat',
-							'background-size': '100% 100%',
-						});
-						beforeCss('#game-background', {
-							'background-image': gameBgScreen
-						});
-						this.clickedGame = game;
-						let self: VitrineClient = this;
-						let events: any = {
-							click() {
-								(<any>$('#game-cover-component')).animateCss('pulse', 120);
-								if (!self.gameLaunched) {
-									ipcRenderer.send('client.launch-game', self.clickedGame.uuid);
-									self.gameLaunched = true;
-								}
-							}
-						};
-						$(document.body).on(events, '#game-cover-image');
-						$(document.body).on(events, '#cover-play-btn');
+					let gameCover: string = 'url(' + game.details.cover.split('\\').join('\\\\') + ')';
+					let gameBgScreen: string = 'url(' + game.details.backgroundScreen.split('\\').join('\\\\') + ')';
+					$('#game-cover-container').css({
+						'display': 'block'
 					});
+					$('#game-title').html(game.name);
+					$('#game-desc').addClass('game-desc').html(game.details.summary);
+					$('#game-cover-image').css({
+						'background-image': gameCover,
+						'background-repeat': 'no-repeat',
+						'background-size': '100% 100%',
+					});
+					beforeCss('#game-background', {
+						'background-image': gameBgScreen
+					});
+					this.clickedGame = game;
+					let self: VitrineClient = this;
+					let events: any = {
+						click() {
+							(<any>$('#game-cover-component')).animateCss('pulse', 120);
+							if (!self.gameLaunched) {
+								ipcRenderer.send('client.launch-game', self.clickedGame.uuid);
+								self.gameLaunched = true;
+							}
+						}
+					};
+					$(document.body).on(events, '#game-cover-image');
+					$(document.body).on(events, '#cover-play-btn');
 				});
 			});
-		}
-		else {
-			$('li.add-game-link[game-id]').each((index, value) => {
-				$(value).click(() => {
-					let gameId: string = $(value).attr('game-id');
-					ipcRenderer.send('client.add-game', gameId);
-				});
-			});
-		}
+		});
 	}
 }
