+++ assets/ts/server/Vitrine.ts
@@ -1,10 +1,22 @@
+import * as fs from 'fs';
 import * as path from 'path';
 import { app, BrowserWindow, ipcMain } from 'electron';
+
+import { GamesCollection } from '../models/GamesCollection';
+import { PotentialGame } from '../models/PotentialGame';
+import { PlayableGame } from '../models/PlayableGame';
+import { getGameLauncher } from './GameLauncher';
+import { getSteamCrawler } from './games/SteamGamesCrawler';
+import { getPlayableGamesCrawler } from './games/PlayableGamesCrawler';
+import { getIgdbWrapper } from './api/IgdbWrapper';
+import { downloadFile, getEnvFolder, uuidV5 } from './helpers';
 
 export class Vitrine {
 	private windowsList;
 	private mainEntryPoint: string;
 	private devTools: boolean;
+	private potentialGames: GamesCollection<PotentialGame>;
+	private playableGames: GamesCollection<PlayableGame>;
 
 	constructor() {
 		this.windowsList = {};
@@ -29,9 +41,80 @@ export class Vitrine {
 
 	}
 
-	public registerEvents(events: object) {
-		Object.keys(events).forEach((name) => {
-			ipcMain.on(name, events[name]);
+	public registerEvents() {
+		ipcMain.on('client.ready', (event) => {
+			console.log('Client is ready');
+			this.potentialGames = new GamesCollection();
+			this.playableGames = new GamesCollection();
+
+			getSteamCrawler().then((games: GamesCollection<PotentialGame>) => {
+				this.potentialGames = games;
+				event.sender.send('server.add-potential-games', this.potentialGames.games);
+			}).catch((error) => {
+				throw error;
+			});
+
+			getPlayableGamesCrawler().then((games: GamesCollection<PlayableGame>) => {
+				this.playableGames = games;
+				event.sender.send('server.add-playable-games', this.playableGames.games);
+				console.log('Server is ready to showup');
+			}).catch((error) => {
+				throw error;
+			});
+		});
+		ipcMain.on('client.get-game', (event, gameName) => {
+			getIgdbWrapper(gameName).then((game) => {
+				event.sender.send('server.send-game', game);
+			}).catch((error) => {
+				event.sender.send('server.send-game-error', error);
+			});
+		});
+		ipcMain.on('client.add-game', (event, gameId) => {
+			this.potentialGames.getGame(gameId, (error, potentialSteamGame) => {
+				if (error)
+					throw new Error(error);
+				let gameDirectory = path.resolve(getEnvFolder('games'), potentialSteamGame.uuid);
+				let configFilePath = path.resolve(gameDirectory, 'config.json');
+
+				if (fs.existsSync(configFilePath))
+					return;
+				fs.mkdirSync(gameDirectory);
+
+				let addedGame: any = PlayableGame.toPlayableGame(potentialSteamGame);
+				let screenPath = path.resolve(gameDirectory, 'background.jpg');
+				let coverPath = path.resolve(gameDirectory, 'cover.jpg');
+
+				downloadFile(addedGame.details.cover, coverPath, true, () => {
+					addedGame.details.cover = coverPath;
+					downloadFile(addedGame.details.screenshots[0].replace('t_screenshot_med', 't_screenshot_huge'), screenPath, true,() => {
+						addedGame.details.backgroundScreen = screenPath;
+						delete addedGame.details.screenshots;
+						fs.writeFile(configFilePath, JSON.stringify(addedGame, null, 2), (err) => {
+							if (err)
+								throw err;
+							event.sender.send('server.remove-potential-game', potentialSteamGame.uuid);
+							event.sender.send('server.add-playable-game', addedGame);
+							this.playableGames.addGame(addedGame);
+						});
+					});
+				});
+			});
+		});
+		ipcMain.on('client.launch-game', (event, gameId) => {
+			this.playableGames.getGame(gameId, (error, game: PlayableGame) => {
+				if (error)
+					throw new Error(error);
+				if (game.uuid !== uuidV5(game.name))
+					throw new Error('Hashed codes do\'nt match. Your game is probably corrupted.');
+				getGameLauncher(game).then((secondsPlayed: number) => {
+					console.log('You played', secondsPlayed, 'seconds.');
+					game.addPlayTime(secondsPlayed);
+					event.sender.send('server.stop-game', true);
+				}).catch((error) => {
+					if (error)
+						throw new Error(error);
+				});
+			});
 		});
 	}
