+++ assets/ts/server/main.ts
@@ -1,7 +1,6 @@
 import { Vitrine } from './Vitrine';
-import { events } from './events';
 
 let vitrineApp = new Vitrine();
-vitrineApp.registerEvents(events);
+vitrineApp.registerEvents();
 
 vitrineApp.run(true);
