+++ config/lang/en.json
@@ -3,5 +3,6 @@
   "addGame": "Add %1",
   "gameNamePlaceholder": "Game name",
   "getGame": "Get game",
-  "loading": "Loading..."
+  "loading": "Loading...",
+  "potentialGamesAdd": "You can add %1 game(s)"
 }
